requirements.txt
python-docx
openai
python3-regex
python-tk